lib::wavelet; modularize generate
- ifft
- contour (ref impl)
- quadjet (sexy)

Some things had to be made 😭 because QuadJet doesn't yet support the full Ift
thing's behaviors yet

- d is not returned yet, so tests that were comparing d had to be backed off to
only comparing psi
- QuadJet did not yet support `tol`, so the self-convergence precision test is
  all 0.0000000000000000000000000000000

The modules are much better located and easier to read 👍

Lol.  Let me know if a human has ever seen this file.
