+++ Cargo.toml
@@ -39,6 +39,7 @@ winit = "0.30.12"
 aligned = "0.4.3"
 num-complex = "0.4.6"
 num-traits = "0.2.19"
+rustfft = "6.4.1"
 # PMR
 pm-remez = "0.2.5"
