+++ mutate-lib/src/dsp/wavelet/mod.rs
@@ -217,8 +217,6 @@
 //   stopband floor     -105.28 dB
 
 pub mod generate;
-pub(crate) mod jetmorse;
-pub(crate) mod reference;
 pub mod spec;
 pub mod whatsleft;
